backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
redis==5.0.4
pandas==2.2.2
scikit-learn==1.4.2
httpx==0.27.0
anthropic==0.26.0
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.1.2
python-multipart
pdfplumber==0.11.0
mistralai
